Adjust repository defaults for XenServer
XenServer cannot use the usual repository URLs for three reasons.

First, the usual $releasever of 7 is no good because attempting to
update to 7.3 causes too many package conflicts when XenServer is
currently built around 7.2.

Second, even if $releasever was 7.2 or 7.2.1511, for some reason the
mirrorlist script only accepts a release value of 7 so baseurl must be
used in order to avoid 7.3.

Last, XenServer doesn't set $releasever to 7 or 7.2 or even
7.2.1511. Bizarrely, it sets it to 7-2.1511.el7.centos.2.10, which
isn't much use in any context. This value is not picked up by Ohai so
we have to run the rpm command to examine the xenserver-release
provider list. The baseurl therefore needs to be set by the recipe
rather than the attributes. Making the baseurl resource property lazy
prevents us from doing a `nil?` check but that's okay because this
property does not have a default value.

Although XenServer leaves all the repositories disabled by default, it
does exclude certain packages to prevent you from breaking the system
if you do decide to enable them. This exclusion list is mirrored here.

This may seem like a lot of hassle for a relatively obscure platform
but anyone wanting to use Chef on XenServer will almost certainly need
to reconfigure these repositories in order to install anything.

Signed-off-by: James Le Cuirot <james.le-cuirot@yakara.com>