Remove support for XenServer
The XenServer code included in #23 has no way to be tested to ensure it
continues to work. In addition, #35 includes support for pulling in Xen directly
from CentOS.  Overall, this adds complexity to the cookbook which isn't needed.
This is really outside of the scope for this cookbook.

If users want to add this feature back, please include ways to do integration
testing to ensure this works.

Signed-off-by: Lance Albertson <lance@osuosl.org>