Ensure requiring on cookbook for non-RHEL systems does not return an error
This resolves #39 which only happens on non-RHEL systems. This is due to the
fact that we're modifying strings in attributes which gets loaded whenver this
cookbook is a dependancy on non-RHEL systems.

This converts whatever value is presented to a string so that the string methods
we use will still work and not cause an error during compile time. In addition,
I added a simple ChefSpec on Ubuntu to try and catch these issues better in the
future.

Signed-off-by: Lance Albertson <lance@osuosl.org>